Add a target for building the data store
Choose the scheme BuildDataStore and run it in the Sim. Copy the 'open /filepath' command from the debugger console and paste it into a Terminal window and it will open the Documents folder. Copy the CodeSearch.sqlite file from that folder into the Project folder and replace the existing file of the same name.